+++ org.openhab.binding.souliss/src/main/java/org/openhab/binding/souliss/internal/discovery/SoulissGatewayDiscovery.java
@@ -195,7 +195,8 @@ public class SoulissGatewayDiscovery extends AbstractDiscoveryService implements
         if (gw != null) {
             gatewayUID = gw.getThing().getUID();
 
-            if (lastByteGatewayIP == Byte.parseByte(gw.IPAddressOnLAN.split("\\.")[3])) {
+            if (lastByteGatewayIP == (byte) Integer.parseInt(gw.IPAddressOnLAN.split("\\.")[3])) {
+
                 String sNodeId = node + SoulissBindingConstants.UUID_NODE_SLOT_SEPARATOR + slot;
                 switch (typical) {
                     case SoulissBindingProtocolConstants.Souliss_T11:
